+++ examples/typescript-react-example/src/InnerModuleSignature.gen.tsx
@@ -8,6 +8,6 @@ const InnerModuleSignatureBS = require('./InnerModuleSignature.bs');
 // tslint:disable-next-line:interface-over-type-literal
 export type InnerModule_t = string;
 
-export const InnerModule_make: (_1:InnerModule_t) => string = InnerModuleSignatureBS.InnerModule.InnerModule;
+export const InnerModule_make: (_1:InnerModule_t) => string = InnerModuleSignatureBS.InnerModule.make;
 
 export const InnerModule: { make: (_1:InnerModule_t) => string } = InnerModuleSignatureBS.InnerModule