+++ Changes.md
@@ -1,5 +1,6 @@
 # master
 - Fix conversion for nested variants from bs-platform 8 onwards.
+- Fix name of functions emitted when an inner signature is annotated (see https://github.com/reason-association/genType/issues/474).
 
 # 3.38.0
 - Add missing support for annotations in inner module signatures (e.g. `module M : { ... } = { ... }`). See https://github.com/reason-association/genType/issues/469.